# src/openapi_server/models/problem_details.py
# coding: utf-8

"""
    Nncof_EventsSubscription

    Nncof_EventsSubscription Service API.   © 2025, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).   All rights reserved. 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictInt, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
from typing_extensions import Annotated
from openapi_server.models.invalid_param import InvalidParam
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class ProblemDetails(BaseModel):
    """
    Provides additional information in an error response.
    """ # noqa: E501
    type: Optional[StrictStr] = Field(default=None, description="String providing an URI formatted according to RFC 3986.")
    title: Optional[StrictStr] = None
    status: Optional[StrictInt] = None
    detail: Optional[StrictStr] = Field(default=None, description="A human-readable explanation specific to this occurrence of the problem.")
    instance: Optional[StrictStr] = Field(default=None, description="String providing an URI formatted according to RFC 3986.")
    cause: Optional[StrictStr] = Field(default=None, description="A machine-readable application error cause specific to this occurrence of the problem.  This IE should be present and provide application-related error information, if available. ")
    invalid_params: Optional[Annotated[List[InvalidParam], Field(min_length=1)]] = Field(default=None, alias="invalidParams")
    supported_features: Optional[Annotated[str, Field(strict=True)]] = Field(default=None, description="A string used to indicate the features supported by an API that is used as defined in clause  6.6 in 3GPP TS 29.500. The string shall contain a bitmask indicating supported features in  hexadecimal representation Each character in the string shall take a value of \"0\" to \"9\",  \"a\" to \"f\" or \"A\" to \"F\" and shall represent the support of 4 features as described in  table 5.2.2-3. The most significant character representing the highest-numbered features shall  appear first in the string, and the character representing features 1 to 4 shall appear last  in the string. The list of features and their numbering (starting with 1) are defined  separately for each API. If the string contains a lower number of characters than there are  defined features for an API, all features that would be represented by characters that are not  present in the string are not supported. ", alias="supportedFeatures")
    supported_api_versions: Optional[Annotated[List[StrictStr], Field(min_length=1)]] = Field(default=None, alias="supportedApiVersions")
    __properties: ClassVar[List[str]] = ["type", "title", "status", "detail", "instance", "cause", "invalidParams", "supportedFeatures", "supportedApiVersions"]

    @field_validator('supported_features')
    def supported_features_validate_regular_expression(cls, value):
        """Validates the regular expression"""
        if value is None:
            return value

        if not re.match(r"^[A-Fa-f0-9]*$", value):
            raise ValueError(r"must validate the regular expression /^[A-Fa-f0-9]*$/")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ProblemDetails from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in invalid_params (list)
        _items = []
        if self.invalid_params:
            for _item in self.invalid_params:
                if _item:
                    _items.append(_item.to_dict())
            _dict['invalidParams'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ProblemDetails from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "type": obj.get("type"),
            "title": obj.get("title"),
            "status": obj.get("status"),
            "detail": obj.get("detail"),
            "instance": obj.get("instance"),
            "cause": obj.get("cause"),
            "invalidParams": [InvalidParam.from_dict(_item) for _item in obj.get("invalidParams")] if obj.get("invalidParams") is not None else None,
            "supportedFeatures": obj.get("supportedFeatures"),
            "supportedApiVersions": obj.get("supportedApiVersions")
        })
        return _obj